Sandbox: ICP: Add material for slides illustration.
	* jardonnet/test/icp_ref.cc: Update. Mark centers of mass.
	* jardonnet/registration/save.hh: Update. Mark center of mass and
	orientation.
	* jardonnet/registration/power_it.hh: Update. Now works in n-D.
	* jardonnet/registration/quat7.hh: Update. Cosmetic change.

	* jardonnet/TODO: Update.
	* jardonnet/test/Makefile: Fix clean rule.


git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@2032 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/sandbox/jardonnet/registration/quat7.hh
@@ -131,17 +131,16 @@ namespace mln
   {
     //mu_Xk = center map(Ck)
     algebra::vec<P::dim,float> mu_Xk(literal::zero);
-    for (size_t i = 0; i < c_length; ++i)
+    for (unsigned i = 0; i < c_length; ++i)
       {
         algebra::vec<P::dim,float> xki = map(Ck[i]);
         mu_Xk += xki;
       }
     mu_Xk /= c_length;
-
    
     // qR
     algebra::mat<P::dim,P::dim,float> Mk(literal::zero);
-    for (size_t i = 0; i < c_length; ++i)
+    for (unsigned i = 0; i < c_length; ++i)
       {
         algebra::vec<P::dim,float> Ci  = C[i];
         algebra::vec<P::dim,float> Xki = map(Ck[i]);
